feat(pacmak): build all java targets at once (#849)
Improve build speed by building all jsii targets at the same time,
and using a special aggregate POM build for Java.

- Improves speed by not unnecessarily re-reading jsii assemblies
  for each individual build.
- Improves speed by using an aggregate POM to build all Java
  targets (can reuse JVM instances and compilation metadata by Maven).
- Improves wall-clock time by building all languages in parallel.

On my machine, reduces the time to pack the entire CDK repository
from 25min user/15min wall clock to 15min user/5 min wall clock.

+++ packages/jsii-pacmak/test/diff-test.sh
@@ -29,6 +29,11 @@ function assert-generator() {
         rsync -a ${original_expected}/ ${expected}/
     fi
 
+    if [[ -d ${original_expected}/java/target ]]; then
+        echo "An IDE plugin seems to have eagerly tried to compile ${original_expected}/java. Please remove." >&2
+        exit 1
+    fi
+
     # put the real expected tarball instead of the placeholder
     for expected_tarball_placeholder in $(find ${expected} -name "*.tgz" || true); do
         rm -f ${expected_tarball_placeholder}
